refactor: move firestore impl into provider
